Добвлена обработка исключений - обработка исключений для сущностей: Task, Status, - добавлены Exception классы для ответа клинту

+++ SoftlineTestTaskApp.Services/Services/StatusService.cs
@@ -1,5 +1,6 @@
 ﻿using SoftlineTestTaskApp.Domain.Dto;
 using SoftlineTestTaskApp.Domain.Entities;
+using SoftlineTestTaskApp.Domain.Exceptions;
 using SoftlineTestTaskApp.Domain.Repositories;
 using SoftlineTestTaskApp.Domain.Services;
 using System.Collections.Generic;
@@ -28,11 +29,23 @@ namespace SoftlineTestTaskApp.Services.Services
         }
         public async System.Threading.Tasks.Task Delete(int id, CancellationToken cancellationToken = default)
         {
+            var status = await _statusRepository.Get(id);
+
+            if (status == null)
+            {
+                throw new EntityNotFoundException(typeof(Status), nameof(id)); 
+            }
+
             await _statusRepository.Delete(id, cancellationToken);
         }
         public async Task<StatusDto> Get(int id, CancellationToken cancellationToken = default)
         {
             var status = await _statusRepository.Get(id, cancellationToken);
+
+            if (status is null)
+            {
+                throw new EntityNotFoundException(typeof(Status), nameof(id));
+            }
 
             return new StatusDto
             {
@@ -44,7 +57,7 @@ namespace SoftlineTestTaskApp.Services.Services
         {
             var statuses = await _statusRepository.GetAll(cancellationToken);
             var statusesDto = new List<StatusDto>(statuses.Count);
-            foreach (var status in statusesDto)
+            foreach (var status in statuses)
             {
                 statusesDto.Add(new StatusDto
                 {
@@ -57,11 +70,19 @@ namespace SoftlineTestTaskApp.Services.Services
         }
         public async System.Threading.Tasks.Task Update(StatusUpdateRequest statusUpdateRequest, CancellationToken cancellationToken = default)
         {
+            var statusExists = await _statusRepository.Get(statusUpdateRequest.Id, cancellationToken);
+
+            if (statusExists == null)
+            {
+                throw new EntityNotFoundException(typeof(Status), nameof(statusUpdateRequest.Id));
+            }
+
             var statusToUpdate = new Status
             {
                 Id = statusUpdateRequest.Id,
                 Name = statusUpdateRequest.Name
             };
+
             await _statusRepository.Update(statusToUpdate, cancellationToken);
         }
     }
